Move GetClassAlignmentRequirementStatic and inline forwarded method
