+++ src/coreclr/tools/Common/Compiler/CompilerTypeSystemContext.Validation.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using Internal.TypeSystem;
 using Internal.TypeSystem.Ecma;
-using Internal.TypeSystem.Interop;
 using Debug = System.Diagnostics.Debug;
 
 namespace ILCompiler
@@ -373,37 +372,6 @@ namespace ILCompiler
             }
 
             return type;
-        }
-
-        /// <summary>
-        /// Managed implementation of CEEInfo::getClassAlignmentRequirementStatic
-        /// </summary>
-        public static int GetClassAlignmentRequirementStatic(DefType type)
-        {
-            int alignment = type.Context.Target.PointerSize;
-
-            if (type is MetadataType metadataType && !metadataType.IsAutoLayout)
-            {
-                if (metadataType.IsSequentialLayout ||
-                    MarshalUtils.IsBlittableType(metadataType))
-                {
-                    alignment = metadataType.InstanceFieldAlignment.AsInt;
-                }
-            }
-            if (type.Context.Target.SupportsAlign8 &&
-                alignment < 8 && type.RequiresAlign8())
-            {
-                // If the structure contains 64-bit primitive fields and the platform requires 8-byte alignment for
-                // such fields then make sure we return at least 8-byte alignment. Note that it's technically possible
-                // to create unmanaged APIs that take unaligned structures containing such fields and this
-                // unconditional alignment bump would cause us to get the calling convention wrong on platforms such
-                // as ARM. If we see such cases in the future we'd need to add another control (such as an alignment
-                // property for the StructLayout attribute or a marshaling directive attribute for p/invoke arguments)
-                // that allows more precise control. For now we'll go with the likely scenario.
-                alignment = 8;
-            }
-
-            return alignment;
         }
     }
 }
